improvement in the technical part of the head

+++ Assets/_Project/Scripts/Improvements/ImprovementWeaponData.cs
@@ -4,6 +4,6 @@
     {
         public int Damage;
         public int SpeedAttack;
-        public int ReturnInitialAttackPosition;
+        public float ReturnInitialAttackPosition;
     }
 }
